consolidate ingestion assets common check used by all PostgresResource jobs ingestion maintains
