+++ dags/locations/ingestion.py
@@ -1,13 +1,12 @@
 import dagster
 
-from dags import delete_persons_from_trigger_log_job, persons_new_backfill
+from dags import delete_persons_from_trigger_log_job, ingestion_assets, persons_new_backfill
 
 from . import resources
 
 defs = dagster.Definitions(
     assets=[
-        persons_new_backfill.postgres_env_check,
-        delete_persons_from_trigger_log_job.postgres_env_check,
+        ingestion_assets.postgres_env_check,
     ],
     jobs=[
         persons_new_backfill.persons_new_backfill_job,
